requirements.txt
altgraph==0.17.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
customtkinter==5.1.3
cycler==0.11.0
darkdetect==0.8.0
easyocr==1.7.0
filelock==3.12.0
fonttools==4.39.3
idna==3.4
imageio==2.30.0
Jinja2==3.1.2
kiwisolver==1.4.4
langdetect==1.0.9
lazy_loader==0.2
macholib==1.16.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.3
ocrmac==0.1.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyclipper==1.3.0.post4
pygame==2.4.0
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pyobjc-core==9.1.1
pyobjc-framework-Cocoa==9.1.1
pyobjc-framework-CoreML==9.1.1
pyobjc-framework-Quartz==9.1.1
pyobjc-framework-Vision==9.1.1
pyparsing==3.0.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
svgwrite==1.4.3
sympy==1.12
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
